+++ common/listener/listener.go
@@ -121,7 +121,7 @@ func (l *Listener) Start() error {
 		if err != nil {
 			return E.Cause(err, "initialize system proxy")
 		}
-		err = systemProxy.Enable()
+		err = applySystemProxy(l.ctx, systemProxy, true)
 		if err != nil {
 			return E.Cause(err, "set system proxy")
 		}
@@ -134,7 +134,7 @@ func (l *Listener) Close() error {
 	l.shutdown.Store(true)
 	var err error
 	if l.systemProxy != nil && l.systemProxy.IsEnabled() {
-		err = l.systemProxy.Disable()
+		err = applySystemProxy(l.ctx, l.systemProxy, false)
 	}
 	return E.Errors(err, common.Close(
 		l.tcpListener,